server/requirements.txt
django==2.0.4
djangorestframework==3.8.2
psycopg2-binary
coreapi
graphene
graphene_django
djangorestframework_simplejwt
django-filter